+++ turboactivate/__init__.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2013, 2014 Develer S.r.l. (http://www.develer.com/)
+# Copyright 2013, 2014 Develer S.r.l. (https://www.develer.com/)
+# Copyright 2017 wyDay, LLC (https://wyday.com/)
 #
 # Author: Lorenzo Villani <lvillani@develer.com>
 # Author: Riccardo Ferrazzo <rferrazz@develer.com>
+# Author: wyDay, LLC <support@wyday.com>
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -23,63 +25,40 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 
-import sys
-
-from datetime import datetime
-
 from ctypes import pointer, sizeof, c_uint32
 
-from c_wrapper import *
+from turboactivate.c_wrapper import *
 
 #
 # Object oriented interface
 #
 
-
-class GenuineOptions(object):
-
-    """A set of options to use with is_genuine()"""
-
-    FLAG_SKIP_OFFLINE = 0x00000001
-    FLAG_OFFLINE_SHOW_INET_ERR = 0x00000002
-
-    def __init__(self, flags=0, grace_days=0, days_between_checks=0):
-        self._flags = flags
-        self._grace_days = grace_days
-        self._days_between_checks = days_between_checks
-
-    def get_pointer(self):
-        options = GENUINE_OPTIONS(sizeof(GENUINE_OPTIONS()),
-                                  self._flags,
-                                  self._days_between_checks,
-                                  self._grace_days)
-        return pointer(options)
-
-    def flags(self, flags):
-        self._flags = flags
-
-    def grace_days(self, days):
-        """
-        If the call fails because of an internet error,
-        how long, in days, should the grace period last (before
-        returning deactivating and returning TA_FAIL).
-
-        14 days is recommended.
-        """
-        self._grace_days = days
-
-    def days_between_checks(self, days):
-        """How often to contact the LimeLM servers for validation. 90 days recommended."""
-        self._days_between_checks = days
+class IsGenuineResult:
+    Genuine, GenuineFeaturesChanged, NotGenuine, NotGenuineInVM, InternetError = range(5)
 
 
 class TurboActivate(object):
 
-    def __init__(self, dat_file, guid, use_trial=False, library_folder="", mode=TA_USER):
+    def __init__(self, dat_file, guid, library_folder="", mode=TA_USER):
         self._lib = load_library(library_folder)
         self._set_restype()
 
-        self.set_current_product(dat_file, guid, use_trial=use_trial, mode=mode)
+        self._mode = mode
+        self._dat_file = wstr(dat_file)
+
+        try:
+            self._lib.TA_PDetsFromPath(self._dat_file)
+        except TurboActivateFailError:
+            # The dat file is already loaded
+            pass
+
+        self._handle = self._lib.TA_GetHandle(wstr(guid))
+
+        # if the handle is still unset then immediately throw an exception
+        # telling the user that they need to actually load the correct
+        # TurboActivate.dat and/or use the correct GUID for the TurboActivate.dat
+        if self._handle == 0:
+            raise TurboActivateDatFileError()
 
     #
     # Public
@@ -87,38 +66,16 @@ class TurboActivate(object):
 
     # Product management
 
-    def set_current_product(self, dat_file, guid, use_trial=False, mode=TA_USER):
-        """
-        This functions allows you to use licensing for multiple products within
-        the same running process.
-        """
-        self._mode = mode
-        self._dat_file = wstr(dat_file)
-        self._guid = wstr(guid)
+    def use_trial(self, verified=True):
 
-        try:
-            self._lib.PDetsFromPath(self._dat_file)
-        except TurboActivateFailError:
-            # The dat file is already loaded
-            pass
+        flags = TA_VERIFIED_TRIAL | self._mode if verified else TA_UNVERIFIED_TRIAL | self._mode
 
-        self._lib.SetCurrentProduct(self._guid)
+        self._lib.TA_UseTrial(self._handle, flags, None)
 
-        if use_trial:
-            self._lib.UseTrial(self._mode)
-
-    def current_product(self):
-        """Gets the "current product" previously set by set_current_product()."""
-        buf_size = 128
-        buf = wbuf(buf_size)
-
-        self._lib.GetCurrentProduct(buf, buf_size)
-
-        return buf.value
 
     # Product key
 
-    def product_key(self):
+    def get_pkey(self):
         """
         Gets the stored product key. NOTE: if you want to check if a product key is valid
         simply call is_product_key_valid().
@@ -127,33 +84,23 @@ class TurboActivate(object):
         buf = wbuf(buf_size)
 
         try:
-            self._lib.GetPKey(buf, buf_size)
+            self._lib.TA_GetPKey(self._handle, buf, buf_size)
 
             return buf.value
         except TurboActivateProductKeyError as e:
             return None
 
-    def set_product_key(self, product_key):
+    def check_and_save_pkey(self, product_key):
         """Checks and saves the product key."""
-        self._lib.CheckAndSavePKey(wstr(product_key), self._mode)
-
-    def blacklists_keys(self, keys_list=[]):
-        """
-        Blacklists keys so they are no longer valid. Use "BlackListKeys" only if
-        you're using the "Serial-only plan" in LimeLM. Otherwise revoke keys.
-        """
-        arr = (wstr * len(keys_list))()
-        arr[:] = keys_list
-
-        self._lib.BlackListKeys(arr, len(arr))
+        self._lib.TA_CheckAndSavePKey(self._handle, wstr(product_key), self._mode)
 
     def is_product_key_valid(self):
         """
         Checks if the product key installed for this product is valid. This does NOT check if
-        the product key is activated or genuine. Use is_activated() and is_genuine() instead.
+        the product key is activated or genuine. Use is_activated() and is_genuine_ex() instead.
         """
         try:
-            self._lib.IsProductKeyValid(self._guid)
+            self._lib.TA_IsProductKeyValid(self._handle)
 
             return True
         except TurboActivateError:
@@ -161,60 +108,98 @@ class TurboActivate(object):
 
     # Activation status
 
-    def deactivate(self, erase_p_key=True, deactivation_request_file=""):
+    def deactivate(self, erase_p_key=False):
         """
-        Deactivates the product on this computer. Set erase_p_key to True to erase the stored
-        product key, False to keep the product key around. If you're using deactivate to let
-        a user move between computers it's almost always best to *not* erase the product
-        key. This way you can just use activate() when the user wants to reactivate
-        instead of forcing the user to re-enter their product key over-and-over again.
-        If deactivation_request_file is specified, then it gets the "deactivation request"
-        file for offline deactivation.
+       Deactivates the product on this computer. Set erasePkey to 1 to erase the stored
+       product key, 0 to keep the product key around. If you're using deactivate to let
+       a user move between computers it's almost always best to *not* erase the product
+       key. This way you can just use TA_Activate() when the user wants to reactivate
+       instead of forcing the user to re-enter their product key over-and-over again.
         """
-        e = '1' if erase_p_key else '0'
-        fn = self._lib.DeactivationRequestToFile if deactivation_request_file else self._lib.Deactivate
-        args = [wstr(deactivation_request_file)] if deactivation_request_file else []
 
-        args.append(e)
+        if erase_p_key is True:
+            args = 1
+        else:
+            args = 0
 
         try:
-            fn(*args)
+            self._lib.TA_Deactivate(self._handle, args)
         except TurboActivateNotActivatedError:
             return
 
-    def activate(self, extra_data="", activation_request_file=""):
+    def deactivation_request_to_file(self, filename, erase_p_key=False):
+        """
+        Get the "deactivation request" file for offline deactivation. Set erase_p_key to
+        True to erase the stored product key, False to keep the product key around. If
+        you're using deactivate to let a user move between computers it's almost always
+        best to *not* erase the product key. This way you can just use
+        activation_request_to_file() when the user wants to reactivate instead of forcing
+        the user to re-enter their product key over-and-over again.
+        """
+
+        args = [wstr(filename)]
+
+        if erase_p_key is True:
+            args.append(1)
+        else:
+            args.append(0)
+
+        try:
+            self._lib.TA_DeactivationRequestToFile(self._handle, *args)
+        except TurboActivateNotActivatedError:
+            return
+
+    def activate(self, extra_data=""):
         """
         Activates the product on this computer. You must call set_product_key()
         with a valid product key or have used the TurboActivate wizard sometime
         before calling this function.
-        If activation_request_file is specified, then it gets the "activation request"
-        file for offline activation.
         """
-        if self.is_activated():
-            return False
-
-        fn = self._lib.ActivationRequestToFile if activation_request_file else self._lib.Activate
-        args = [wstr(activation_request_file)] if activation_request_file else []
 
         if extra_data:
-            fn = self._lib.ActivationRequestToFileEx if activation_request_file else self._lib.ActivateEx
             options = ACTIVATE_OPTIONS(sizeof(ACTIVATE_OPTIONS()),
-                                       wstr(extra_data))
-            args.append(pointer(options))
+                                      wstr(extra_data))
+
+            args = pointer(options)
+        else:
+            args = None
 
         try:
-            fn(*args)
+            self._lib.TA_Activate(self._handle, args)
 
             return True
         except TurboActivateError as e:
-            if not activation_request_file:
-                self.deactivate(True)
+            raise e
 
+    def activation_request_to_file(self, filename, extra_data=""):
+        """
+        Get the "activation request" file for offline activation. You must call
+        set_product_key() with a valid product key or have used the TurboActivate
+        Wizard sometime before calling this function.
+        """
+        args = [wstr(filename)]
+
+        if extra_data:
+            options = ACTIVATE_OPTIONS(sizeof(ACTIVATE_OPTIONS()),
+                                      wstr(extra_data))
+
+            args.append(pointer(options))
+        else:
+            args.append(None)
+
+        try:
+            self._lib.TA_ActivationRequestToFile(self._handle, *args)
+
+            return True
+        except TurboActivateError as e:
             raise e
 
     def activate_from_file(self, filename):
         """Activate from the "activation response" file for offline activation."""
-        self._lib.ActivateFromFile(wstr(filename))
+        try:
+            self._lib.TA_ActivateFromFile(self._handle, wstr(filename))
+        except TurboActivateFailError as e:
+            raise e
 
     def get_extra_data(self):
         """Gets the extra data you passed in using activate()"""
@@ -222,7 +207,7 @@ class TurboActivate(object):
         buf = wbuf(buf_size)
 
         try:
-            self._lib.GetExtraData(buf, buf_size)
+            self._lib.TA_GetExtraData(self._handle, buf, buf_size)
 
             return buf.value
         except TurboActivateFailError:
@@ -230,12 +215,16 @@ class TurboActivate(object):
 
     def is_activated(self):
         """ Checks whether the computer has been activated."""
-        try:
-            self._lib.IsActivated(self._guid)
 
+        ret = self._lib.TA_IsActivated(self._handle)
+
+        if ret == TA_OK:
             return True
-        except TurboActivateError:
+        elif ret == TA_FAIL:
             return False
+
+        # raise an error on all other return codes
+        validate_result(ret)
 
     # Features
 
@@ -244,70 +233,100 @@ class TurboActivate(object):
 
     def get_feature_value(self, name):
         """Gets the value of a feature."""
-        buf_size = self._lib.GetFeatureValue(wstr(name), 0, 0)
+        buf_size = self._lib.TA_GetFeatureValue(self._handle, wstr(name), 0, 0)
         buf = wbuf(buf_size)
 
-        self._lib.GetFeatureValue(wstr(name), buf, buf_size)
+        self._lib.TA_GetFeatureValue(self._handle, wstr(name), buf, buf_size)
 
         return buf.value
 
     # Genuine
 
-    def is_genuine(self, options=None):
+    def is_genuine(self):
         """
         Checks whether the computer is genuinely activated by verifying with the LimeLM servers.
         If reactivation is needed then it will do this as well.
-        Optionally you can pass a GenuineOptions object to specify more details
         """
-        fn = self._lib.IsGenuine
-        args = [self._guid]
+        ret = self._lib.TA_IsGenuine(self._handle)
 
-        if options:
-            fn = self._lib.IsGenuineEx
+        if ret == TA_OK:
+            return IsGenuineResult.Genuine
+        elif ret in {TA_FAIL, TA_E_REVOKED, TA_E_ACTIVATE}:
+            return IsGenuineResult.NotGenuine
+        elif ret == TA_E_INET:
+            return IsGenuineResult.InternetError
+        elif ret == TA_E_IN_VM:
+            return IsGenuineResult.NotGenuineInVM
+        elif ret == TA_E_FEATURES_CHANGED:
+            return IsGenuineResult.GenuineFeaturesChanged
 
-            args.append(options.get_pointer())
+        validate_result(ret)
 
-        try:
-            fn(*args)
+    # IsGenuineEx
 
-            return True
-        except TurboActivateFeaturesChangedError:
-            return True
+    def is_genuine_ex(self, days_between_checks, grace_days_on_inet_err, skip_offline = False, offline_show_inet_err = False):
+        """
+        Checks whether the computer is genuinely activated by verifying with the LimeLM servers.
+        If reactivation is needed then it will do this as well.
+        """
+        flags = 0
+
+        if skip_offline:
+            flags = TA_SKIP_OFFLINE
+
+        if offline_show_inet_err:
+            flags = flags | TA_OFFLINE_SHOW_INET_ERR
+
+        options = GENUINE_OPTIONS(sizeof(GENUINE_OPTIONS()),
+                                  flags,
+                                  days_between_checks,
+                                  grace_days_on_inet_err)
+
+        ret = self._lib.TA_IsGenuineEx(self._handle, pointer(options))
+
+        if ret == TA_OK:
+            return IsGenuineResult.Genuine
+        elif ret in {TA_FAIL, TA_E_REVOKED, TA_E_ACTIVATE}:
+            return IsGenuineResult.NotGenuine
+        elif ret in {TA_E_INET, TA_E_INET_DELAYED}:
+            return IsGenuineResult.InternetError
+        elif ret == TA_E_IN_VM:
+            return IsGenuineResult.NotGenuineInVM
+        elif ret == TA_E_FEATURES_CHANGED:
+            return IsGenuineResult.GenuineFeaturesChanged
+
+        validate_result(ret)
 
     # Trial
 
-    def trial_days_remaining(self):
+    def trial_days_remaining(self, verified=True):
         """
         Get the number of trial days remaining.
         0 days if the trial has expired or has been tampered with
         (1 day means *at most* 1 day, that is it could be 30 seconds)
 
-        You must have instantiated TurboActivate with the use_trial
-        flag to use this function
+        You must have called "use_trial" o use this function
         """
+        flags = TA_VERIFIED_TRIAL | self._mode if verified else TA_UNVERIFIED_TRIAL | self._mode
         days = c_uint32(0)
 
-        self._lib.TrialDaysRemaining(self._guid, pointer(days))
+        self._lib.TA_TrialDaysRemaining(self._handle, flags, pointer(days))
 
         return days.value
 
     def extend_trial(self, extension_code):
         """Extends the trial using a trial extension created in LimeLM."""
-        self._lib.ExtendTrial(wstr(extension_code))
+        self._lib.TA_ExtendTrial(self._handle, self._mode, wstr(extension_code))
 
     # Utils
 
-    def is_date_valid(self, date=None):
+    def is_date_valid(self, date):
         """
         Check if the date is valid
         """
-        if not date:
-            to_check = datetime.utcnow().strftime("%Y-%m-%d %H-%M-%S")
-        else:
-            to_check = date
 
         try:
-            self._lib.IsDateValid(wstr(to_check), TA_HAS_NOT_EXPIRED)
+            self._lib.TA_IsDateValid(self._handle, wstr(date), TA_HAS_NOT_EXPIRED)
 
             return True
         except TurboActivateFlagsError as e:
@@ -315,7 +334,7 @@ class TurboActivate(object):
         except TurboActivateError:
             return False
 
-    def set_custom_path(self, path):
+    def set_custom_act_data_path(self, path):
         """
         This function allows you to set a custom folder to store the activation
         data files. For normal use we do not recommend you use this function.
@@ -331,13 +350,9 @@ class TurboActivate(object):
 
         The directory you pass in must already exist. And the process using TurboActivate
         must have permission to create, write, and delete files in that directory.
-
-        On linux it is not available
         """
-        if sys.platform.startswith('linux'):
-            raise RuntimeError("set_custom_path is not available under linux")
 
-        self._lib.SetCustomActDataPath(wstr(path))
+        self._lib.TA_SetCustomActDataPath(wstr(path))
 
     def set_custom_proxy(self, address):
         """
@@ -351,34 +366,22 @@ class TurboActivate(object):
 
         If the port is not specified, TurboActivate will default to using port 1080 for proxies.
         """
-        self._lib.SetCustomProxy(wstr(address))
+        self._lib.TA_SetCustomProxy(wstr(address))
 
     def _set_restype(self):
-        self._lib.PDetsFromPath.restype = validate_result
-        self._lib.SetCurrentProduct.restype = validate_result
-        self._lib.UseTrial.restype = validate_result
-        self._lib.GetCurrentProduct.restype = validate_result
-        self._lib.GetPKey.restype = validate_result
-        self._lib.CheckAndSavePKey.restype = validate_result
-        self._lib.BlackListKeys.restype = validate_result
-        self._lib.IsProductKeyValid.restype = validate_result
-        self._lib.DeactivationRequestToFile.restype = validate_result
-        self._lib.Deactivate.restype = validate_result
-        self._lib.Activate.restype = validate_result
-        self._lib.ActivationRequestToFile.restype = validate_result
-        self._lib.ActivationRequestToFileEx.restype = validate_result
-        self._lib.ActivateEx.restype = validate_result
-        self._lib.ActivateFromFile.restype = validate_result
-        self._lib.GetExtraData.restype = validate_result
-        self._lib.IsActivated.restype = validate_result
-        self._lib.IsGenuine.restype = validate_result
-        self._lib.IsGenuineEx.restype = validate_result
-        self._lib.TrialDaysRemaining.restype = validate_result
-        self._lib.ExtendTrial.restype = validate_result
-        self._lib.IsDateValid.restype = validate_result
-        self._lib.IsDateValid.restype = validate_result
-        self._lib.SetCustomProxy.restype = validate_result
-
-        # SetCustomActDataPath is not defined under linux
-        if not sys.platform.startswith('linux'):
-            self._lib.SetCustomActDataPath.restype = validate_result
+        self._lib.TA_PDetsFromPath.restype = validate_result
+        self._lib.TA_UseTrial.restype = validate_result
+        self._lib.TA_GetPKey.restype = validate_result
+        self._lib.TA_CheckAndSavePKey.restype = validate_result
+        self._lib.TA_IsProductKeyValid.restype = validate_result
+        self._lib.TA_DeactivationRequestToFile.restype = validate_result
+        self._lib.TA_Deactivate.restype = validate_result
+        self._lib.TA_Activate.restype = validate_result
+        self._lib.TA_ActivationRequestToFile.restype = validate_result
+        self._lib.TA_ActivateFromFile.restype = validate_result
+        self._lib.TA_GetExtraData.restype = validate_result
+        self._lib.TA_TrialDaysRemaining.restype = validate_result
+        self._lib.TA_ExtendTrial.restype = validate_result
+        self._lib.TA_IsDateValid.restype = validate_result
+        self._lib.TA_SetCustomProxy.restype = validate_result
+        self._lib.TA_SetCustomActDataPath.restype = validate_result