+++ turboactivate/c_wrapper.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2013, 2014 Develer S.r.l. (http://www.develer.com/)
+# Copyright 2013, 2014 Develer S.r.l. (https://www.develer.com/)
+# Copyright 2017 wyDay, LLC (https://wyday.com/)
 #
 # Author: Lorenzo Villani <lvillani@develer.com>
 # Author: Riccardo Ferrazzo <rferrazz@develer.com>
+# Author: wyDay, LLC <support@wyday.com>
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -62,12 +64,23 @@ TA_E_INSUFFICIENT_BUFFER = 0x0000000E
 TA_E_PERMISSION = 0x0000000F
 TA_E_INVALID_FLAGS = 0x00000010
 TA_E_IN_VM = 0x00000011
+TA_E_IN_SANDBOX = 0x00000022
 TA_E_EDATA_LONG = 0x00000012
 TA_E_INVALID_ARGS = 0x00000013
 TA_E_KEY_FOR_TURBOFLOAT = 0x00000014
 TA_E_INET_DELAYED = 0x00000015
 TA_E_FEATURES_CHANGED = 0x00000016
 TA_E_ANDROID_NOT_INIT = 0x00000017
+TA_E_NO_MORE_DEACTIVATIONS = 0x00000018
+TA_E_ACCOUNT_CANCELED = 0x00000019
+TA_E_ALREADY_ACTIVATED = 0x0000001A
+TA_E_INVALID_HANDLE = 0x0000001B
+TA_E_ENABLE_NETWORK_ADAPTERS = 0x0000001C
+TA_E_ALREADY_VERIFIED_TRIAL = 0x0000001D
+TA_E_TRIAL_EXPIRED = 0x0000001E
+TA_E_MUST_SPECIFY_TRIAL_TYPE = 0x0000001F
+TA_E_MUST_USE_TRIAL = 0x00000020
+TA_E_NO_MORE_TRIALS_ALLOWED = 0x00000021
 
 # Flags for the UseTrial() and CheckAndSavePKey() functions.
 
@@ -76,7 +89,6 @@ TA_USER = 0x00000002
 
 # Flags for the IsGeninueEx() function.
 
-TA_SKIP_OFFLINE = 0x00000001
 """
 If the user activated using offline activation
 (ActivateRequestToFile(), ActivateFromFile() ), then with this
@@ -92,8 +104,8 @@ this flag in tandem with TA_OFFLINE_SHOW_INET_ERR.
 If the user activated using online activation then this flag
 is ignored.
 """
+TA_SKIP_OFFLINE = 0x00000001
 
-TA_OFFLINE_SHOW_INET_ERR = 0x00000002
 """
 If the user activated using offline activation, and you're
 using this flag in tandem with TA_SKIP_OFFLINE, then IsGenuineEx()
@@ -102,13 +114,38 @@ will return TA_E_INET on internet failure instead of TA_OK.
 If the user activated using online activation then this flag
 is ignored.
 """
+TA_OFFLINE_SHOW_INET_ERR = 0x00000002
 
-TA_DISALLOW_VM = 0x00000004
 """
 Use the TA_DISALLOW_VM in UseTrial() to disallow trials in virtual machines.
 If you use this flag in UseTrial() and the customer's machine is a Virtual
 Machine, then UseTrial() will return TA_E_IN_VM.
 """
+TA_DISALLOW_VM = 0x00000004
+
+"""
+Use the TA_DISALLOW_SANDBOX flag to explicitly forbid sandbox type applications
+(e.g. Sandboxie, Docker, etc.) that are used to get around trials or extend them
+forever.
+"""
+TA_DISALLOW_SANDBOX = 0x00000008
+
+"""
+Use this flag in TA_UseTrial() to tell TurboActivate to use client-side
+unverified trials. For more information about verified vs. unverified trials,
+see here: https://wyday.com/limelm/help/trials/
+Note: unverified trials are unsecured and can be reset by malicious customers.
+"""
+TA_UNVERIFIED_TRIAL = 0x00000010
+
+
+"""
+Use the TA_VERIFIED_TRIAL flag to use verified trials instead
+of unverified trials. This means the trial is locked to a particular computer.
+The customer can't reset the trial.
+"""
+TA_VERIFIED_TRIAL = 0x00000020
+
 
 # Flags for the is_date_valid() Function
 
@@ -132,13 +169,14 @@ class ACTIVATE_OPTIONS(Structure):
 
 
 def load_library(path):
-    LIBRARIES = {
-        'linux2': ospath.join(path, 'libTurboActivate.so'),
-        'darwin': ospath.join(path, 'libTurboActivate.dylib'),
-        'win32': ospath.join(path, 'TurboActivate.dll'),
-    }
 
-    return cdll.LoadLibrary(LIBRARIES[sys.platform])
+    if sys.platform == 'win32':
+        return cdll.LoadLibrary(ospath.join(path, 'TurboActivate.dll'))
+    elif sys.platform == 'darwin':
+        return cdll.LoadLibrary(ospath.join(path, 'libTurboActivate.dylib'))
+
+    # else: linux, bsd, etc.
+    return cdll.LoadLibrary(ospath.join(path, 'libTurboActivate.so'))
 
 
 def validate_result(return_code):
@@ -149,8 +187,6 @@ def validate_result(return_code):
     # Raise an exception type appropriate for the kind of error
     if return_code == TA_FAIL:
         raise TurboActivateFailError()
-    elif return_code == TA_E_FEATURES_CHANGED:
-        raise TurboActivateFeaturesChangedError()
     elif return_code == TA_E_PDETS:
         raise TurboActivateDatFileError()
     elif return_code == TA_E_EDATA_LONG:
@@ -181,9 +217,31 @@ def validate_result(return_code):
         raise TurboActivateConnectionDelayedError()
     elif return_code == TA_E_PERMISSION:
         raise TurboActivatePermissionError()
+    elif return_code == TA_E_NO_MORE_DEACTIVATIONS:
+        raise TurboActivateNoMoreDeactivationsError()
+    elif return_code == TA_E_ACCOUNT_CANCELED:
+        raise TurboActivateAccountCanceledError()
+    elif return_code == TA_E_INVALID_HANDLE:
+        raise TurboActivateInvalidHandleError()
+    elif return_code == TA_E_ALREADY_ACTIVATED:
+        raise TurboActivateAlreadyActivatedError()
+    elif return_code == TA_E_ENABLE_NETWORK_ADAPTERS:
+        raise TurboActivateEnableNetworkAdaptersError()
+    elif return_code == TA_E_ALREADY_VERIFIED_TRIAL:
+        raise TurboActivateAlreadyVerifiedTrialError()
+    elif return_code == TA_E_TRIAL_EXPIRED:
+        raise TurboActivateTrialExpiredError()
+    elif return_code == TA_E_MUST_SPECIFY_TRIAL_TYPE:
+        raise TurboActivateMustSpecifyTrialTypeError()
+    elif return_code == TA_E_MUST_USE_TRIAL:
+        raise TurboActivateMustUseTrialError()
+    elif return_code == TA_E_NO_MORE_TRIALS_ALLOWED:
+        raise TurboActivateNoMoreTrialsError()
+    elif return_code == TA_E_INVALID_ARGS:
+        raise TurboActivateInvalidArgsError()
 
     # Otherwise bail out and raise a generic exception
-    raise TurboActivateError()
+    raise TurboActivateError(return_code)
 
 
 #
@@ -291,16 +349,6 @@ class TurboActivatePermissionError(TurboActivateError):
     pass
 
 
-class TurboActivateFeaturesChangedError(TurboActivateError):
-
-    """
-    If IsGenuine() or IsGenuineEx() reactivated and the features
-    have changed, then this will be the return code. Treat this
-    as a success.
-    """
-    pass
-
-
 class TurboActivateDatFileError(TurboActivateError):
 
     """The product details file "TurboActivate.dat" failed to load."""
@@ -331,5 +379,98 @@ class TurboActivateConnectionDelayedError(TurboActivateError):
     is_genuine() previously had a TA_E_INET error, and instead
     of hammering the end-user's network, is_genuine() is waiting
     5 hours before rechecking on the network.
+    """
+    pass
+
+
+class TurboActivateNoMoreDeactivationsError(TurboActivateError):
+    """
+    This product key had a limited number of allowed deactivations.
+    No more deactivations are allowed for the product key. This product
+    is still activated on this computer.
+    """
+    pass
+
+
+class TurboActivateAccountCanceledError(TurboActivateError):
+    """
+    Can't activate or start a verified trial because the LimeLM
+    account is cancelled.
+    """
+    pass
+
+
+class TurboActivateAlreadyActivatedError(TurboActivateError):
+    """
+    You can't use a product key because your app is already activated
+    with a product key. To use a new product key, then first deactivate using
+    either the TA_Deactivate() or TA_DeactivationRequestToFile().
+    """
+    pass
+
+
+class TurboActivateInvalidHandleError(TurboActivateError):
+    """
+    The handle is not valid. To get a handle use TA_GetHandle().
+    """
+    pass
+
+
+class TurboActivateEnableNetworkAdaptersError(TurboActivateError):
+    """
+    There are network adapters on the system that are disabled and
+    TurboActivate couldn't read their hardware properties (even after trying
+    and failing to enable the adapters automatically). Enable the network adapters,
+    re-run the function, and TurboActivate will be able to "remember" the adapters
+    even if the adapters are disabled in the future.
+
+    Note:   The network adapters do not need an active Internet connections. They just
+            need to not be disabled. Whether they are or are not connected to the
+            internet/intranet is not important and does not affect this error code at all.
+
+
+    On Linux you'll get this error if you don't have any real network adapters attached.
+    For example if you have no "eth[x]", "wlan[x]", "en[x]", "wl[x]", "ww[x]", or "sl[x]"
+    network interface devices.
+
+    See: https://wyday.com/limelm/help/faq/#disabled-adapters
+    """
+    pass
+
+
+class TurboActivateAlreadyVerifiedTrialError(TurboActivateError):
+    """
+    The trial is already a verified trial. You need to use the "TA_VERIFIED_TRIAL"
+    flag. Can't "downgrade" a verified trial to an unverified trial.
+    """
+    pass
+
+
+class TurboActivateMustSpecifyTrialTypeError(TurboActivateError):
+    """
+    You must specify the trial type (TA_UNVERIFIED_TRIAL or TA_VERIFIED_TRIAL).
+    And you can't use both flags. Choose one or the other. We recommend TA_VERIFIED_TRIAL.
+    """
+    pass
+
+
+class TurboActivateMustUseTrialError(TurboActivateError):
+    """
+    You must call TA_UseTrial() before you can get the number of trial days remaining.
+    """
+    pass
+
+
+class TurboActivateNoMoreTrialsError(TurboActivateError):
+    """
+    In the LimeLM account either the trial days is set to 0, OR the account is set
+    to not auto-upgrade and thus no more verified trials can be made.
+    """
+    pass
+
+
+class TurboActivateInvalidArgsError(TurboActivateError):
+    """
+    The arguments passed to the function are invalid. Double check your logic.
     """
     pass