+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project are documented in this file.
 
+## 4.1.1 - 2017-07-XX
+
+* Works with latest version of TurboActivate (4.1.x) and newer.
+* Add support for Python 3.x+.
+* Completely re-vamp the example app to be more "copy-paste" ready for existing Python apps.
+* Handle an incorrect VersionGUID immediately in the TurboActivate constructor.
+* Remove redundant "GenuineOptions" class (nearly identical structure to GENUINE_OPTIONS).
+* `turboactivate.is_genuine` is split into 2 separate functions (for good reason --
+  they should be used in completely separate contexts)
+* `is_genuine` and `is_genuine_ex` now return an IsGenuineResult instead of raising exceptions.
+* Remove useless helper function, and put it in the TurboActivate constructor.
+* Make the Python TurboActivate class match other object-oriented examples.
+* Remove unused imports in the example and in the TurboActivate class.
+* Change the default for `ta.deactivate()` to *not* delete the product key.
+* Split `ta.activate()` in 2 separate functions (online and offline activation).
+* Add ability to pass [extra data](https://wyday.com/limelm/help/extra-data/) to `ta.activate()`
+* Modify `ta.is_activated()` to raise errors when it needs to.
+* Modify `ta.is_date_valid()` to require a passed in date/time value.
+* Fix Linux support in Python 3.3+ (no longer uses the "linux2" identifier --
+  now uses "linux"). Also add support for BSD, etc. and remove old cruft.
+* Make a first class language integration with LimeLM (meaning supported with
+  the help of wyDay).
+* Changed the version scheme to match that of the TurboActivate API (it will now match, or slightly trail the TurboActivate API).
+
+
 ## 1.0.4 - 2016-01-27
 
 ### Changed